switch the doc of the switch to doc.rs
